# backend/openapi.yaml
openapi: 3.0.0
info:
  title: PropertyPilot Backend API
  description: Real estate investment analysis platform with financial calculations, property management, and reporting.
  version: 1.0.0
  contact:
    name: PropertyPilot Support
    email: support@propertypilot.com
  license:
    name: MIT
servers:
  - url: https://api.propertypilot.com
    description: Production server
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/upload:
    post:
      tags:
        - Uploads
      summary: Upload single file
      description: Upload a single file to property or renovation
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max 50MB)
                propertyId:
                  type: string
                  format: uuid
                  description: Associated property ID (optional)
                renovationId:
                  type: string
                  format: uuid
                  description: Associated renovation ID (optional)
              required:
                - file
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      file:
                        $ref: '#/components/schemas/FileUpload'
  /api/upload/batch:
    post:
      tags:
        - Uploads
      summary: Upload multiple files
      description: Upload multiple files at once
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to upload
                propertyId:
                  type: string
                  format: uuid
              required:
                - files
      responses:
        '201':
          description: Files uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      files:
                        type: array
                        items:
                          $ref: '#/components/schemas/FileUpload'
                      count:
                        type: integer
  /api/uploads/{fileId}:
    delete:
      tags:
        - Uploads
      summary: Delete file
      description: Delete an uploaded file
      security:
        - SessionAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: File deleted successfully
...
  /api/user/me:
    get:
      tags:
        - User
      summary: Get user profile
      description: Get the current user's profile information
      security:
        - SessionAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'

    put:
      tags:
        - User
      summary: Update profile
      description: Update user profile information
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                avatar:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
  /api/user/preferences:
    get:
      tags:
        - User
      summary: Get preferences
      description: Get user preferences
      security:
        - SessionAuth: []
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      preferences:
                        $ref: '#/components/schemas/UserPreferences'

    put:
      tags:
        - User
      summary: Update preferences
      description: Update user preferences
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      preferences:
                        $ref: '#/components/schemas/UserPreferences'
  /api/user/export-data:
    get:
      tags:
        - User
      summary: Export user data
      description: Export all user data (GDPR compliance)
      security:
        - SessionAuth: []
      responses:
        '200':
          description: User data exported
          content:
            application/json:
              schema:
                type: object
  /api/user/delete-account:
    post:
      tags:
        - User
      summary: Delete account
      description: Delete user account and all associated data (GDPR compliance, irreversible)
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: User password for confirmation
              required:
                - password
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Account deleted successfully. All your data has been removed.
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication via httpOnly cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FileUpload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
          nullable: true
        renovationId:
          type: string
          format: uuid
          nullable: true
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        s3Key:
          type: string
        signedUrl:
          type: string
          format: uri
        signedUrlExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    UserPreferences:
      type: object
      properties:
        currency:
          type: string
          default: USD
        timezone:
          type: string
          default: America/New_York
        propertyTypes:
          type: array
          items:
            type: string
            enum: [rent, flip, airbnb]
          default: [rent, flip, airbnb]
        propertyStatuses:
          type: array
          items:
            type: string
          default: [lead, analyzing, offer, under_contract, owned]
...